+++ libs/contracts/http/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    id("java-library")
+    id("org.springframework.boot") apply false
+    id("io.spring.dependency-management")
+}
+
+tasks.bootJar {
+    enabled = false
+}
+
+tasks.jar {
+    enabled = true
+}
+
+dependencies {
+    // Testing
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+}
