+++ libs/security/context/build.gradle.kts
@@ -1,0 +1,32 @@
+plugins {
+    id("java-library")
+    id("org.springframework.boot") apply false
+    id("io.spring.dependency-management")
+}
+
+dependencies {
+    api(project(":libs:contracts:http"))
+    api(project(":libs:core:exception"))
+
+    // Servlet API (for AuthContextCleanupFilter)
+    compileOnly("org.springframework.boot:spring-boot-starter-web")
+    implementation("org.springframework.boot:spring-boot-autoconfigure")
+    annotationProcessor("org.springframework.boot:spring-boot-configuration-processor")
+
+    // Lombok
+    compileOnly("org.projectlombok:lombok")
+    annotationProcessor("org.projectlombok:lombok")
+
+    // Testing
+    testImplementation("org.springframework.boot:spring-boot-starter-web")
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+}
+
+tasks.bootJar {
+    enabled = false
+}
+
+tasks.jar {
+    enabled = true
+}
