+++ libs/security/context/src/main/java/com/example/security/context/HeaderSecurityException.java
@@ -1,0 +1,15 @@
+package com.example.security.context;
+
+import com.example.core.exception.CommonErrorCode;
+import com.example.core.exception.TechnicalException;
+
+public class HeaderSecurityException extends TechnicalException {
+
+    public HeaderSecurityException(String message) {
+        super(CommonErrorCode.UNAUTHORIZED, message);
+    }
+
+    public HeaderSecurityException(String message, Throwable cause) {
+        super(CommonErrorCode.UNAUTHORIZED, message, cause);
+    }
+}
