--- libs/security/core/src/main/java/com/example/security/core/AuthContextCleanupFilter.java
@@ -1,32 +1,0 @@
-package com.example.security.core;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-/**
- * 요청 완료 후 AuthContextHolder의 ThreadLocal을 정리하여
- * 스레드 풀에서의 메모리 누수 및 컨텍스트 유출을 방지한다.
- */
-@Component
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class AuthContextCleanupFilter extends OncePerRequestFilter {
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request,
-                                     HttpServletResponse response,
-                                     FilterChain filterChain) throws ServletException, IOException {
-        try {
-            filterChain.doFilter(request, response);
-        } finally {
-            AuthContextHolder.clear();
-        }
-    }
-}
