--- libs/security/core/src/main/java/com/example/security/core/HeaderSecurityException.java
@@ -1,15 +1,0 @@
-package com.example.security.core;
-
-import com.example.core.exception.CommonErrorCode;
-import com.example.core.exception.TechnicalException;
-
-public class HeaderSecurityException extends TechnicalException {
-
-    public HeaderSecurityException(String message) {
-        super(CommonErrorCode.UNAUTHORIZED, message);
-    }
-
-    public HeaderSecurityException(String message, Throwable cause) {
-        super(CommonErrorCode.UNAUTHORIZED, message, cause);
-    }
-}
