--- libs/security/core/src/main/java/com/example/security/core/encryption/AesGcmEncryptor.java
@@ -1,77 +1,0 @@
-package com.example.security.core.encryption;
-
-import javax.crypto.Cipher;
-import javax.crypto.spec.GCMParameterSpec;
-import javax.crypto.spec.SecretKeySpec;
-import java.nio.charset.StandardCharsets;
-import java.security.SecureRandom;
-import java.util.Arrays;
-import java.util.Base64;
-
-public class AesGcmEncryptor implements Encryptor {
-
-    private static final String ALGORITHM = "AES/GCM/NoPadding";
-    private static final int GCM_IV_LENGTH = 12;
-    private static final int GCM_TAG_LENGTH = 128;
-
-    private final SecretKeySpec keySpec;
-    private final SecureRandom secureRandom;
-
-    public AesGcmEncryptor(String key) {
-        byte[] keyBytes = normalizeKey(key.getBytes(StandardCharsets.UTF_8));
-        this.keySpec = new SecretKeySpec(keyBytes, "AES");
-        this.secureRandom = new SecureRandom();
-    }
-
-    @Override
-    public String encrypt(String plainText) {
-        try {
-            byte[] iv = new byte[GCM_IV_LENGTH];
-            secureRandom.nextBytes(iv);
-
-            Cipher cipher = Cipher.getInstance(ALGORITHM);
-            cipher.init(Cipher.ENCRYPT_MODE, keySpec, new GCMParameterSpec(GCM_TAG_LENGTH, iv));
-
-            byte[] cipherText = cipher.doFinal(plainText.getBytes(StandardCharsets.UTF_8));
-
-            byte[] combined = new byte[iv.length + cipherText.length];
-            System.arraycopy(iv, 0, combined, 0, iv.length);
-            System.arraycopy(cipherText, 0, combined, iv.length, cipherText.length);
-
-            return Base64.getUrlEncoder().withoutPadding().encodeToString(combined);
-        } catch (Exception e) {
-            throw new EncryptionException("암호화에 실패했습니다", e);
-        }
-    }
-
-    @Override
-    public String decrypt(String cipherText) {
-        try {
-            byte[] combined = Base64.getUrlDecoder().decode(cipherText);
-
-            byte[] iv = Arrays.copyOfRange(combined, 0, GCM_IV_LENGTH);
-            byte[] encrypted = Arrays.copyOfRange(combined, GCM_IV_LENGTH, combined.length);
-
-            Cipher cipher = Cipher.getInstance(ALGORITHM);
-            cipher.init(Cipher.DECRYPT_MODE, keySpec, new GCMParameterSpec(GCM_TAG_LENGTH, iv));
-
-            byte[] decrypted = cipher.doFinal(encrypted);
-            return new String(decrypted, StandardCharsets.UTF_8);
-        } catch (Exception e) {
-            throw new EncryptionException("복호화에 실패했습니다", e);
-        }
-    }
-
-    @Override
-    public String getAlgorithm() {
-        return ALGORITHM;
-    }
-
-    private byte[] normalizeKey(byte[] key) {
-        if (key.length == 16 || key.length == 24 || key.length == 32) {
-            return key;
-        }
-        throw new EncryptionException(
-                "AES 키 길이는 16, 24, 32 바이트 중 하나여야 합니다. 현재: " + key.length + " 바이트");
-    }
-}
