--- libs/security/core/src/main/java/com/example/security/core/encryption/EncryptionException.java
@@ -1,15 +1,0 @@
-package com.example.security.core.encryption;
-
-import com.example.core.exception.CommonErrorCode;
-import com.example.core.exception.TechnicalException;
-
-public class EncryptionException extends TechnicalException {
-
-    public EncryptionException(String message) {
-        super(CommonErrorCode.INTERNAL_ERROR, message);
-    }
-
-    public EncryptionException(String message, Throwable cause) {
-        super(CommonErrorCode.INTERNAL_ERROR, message, cause);
-    }
-}
