--- libs/security/core/src/main/java/com/example/security/core/encryption/Encryptor.java
@@ -1,10 +1,0 @@
-package com.example.security.core.encryption;
-
-public interface Encryptor {
-
-    String encrypt(String plainText);
-
-    String decrypt(String cipherText);
-
-    String getAlgorithm();
-}
