--- libs/security/core/src/main/java/com/example/security/core/pii/PiiType.java
@@ -1,11 +1,0 @@
-package com.example.security.core.pii;
-
-public enum PiiType {
-    GENERAL,
-    EMAIL,
-    PHONE,
-    NAME,
-    ADDRESS,
-    CARD_NUMBER,
-    SSN
-}
