+++ libs/security/crypto/build.gradle.kts
@@ -4,22 +4,11 @@ plugins {
     id("io.spring.dependency-management")
 }
 
-tasks.bootJar {
-    enabled = false
-}
-
-tasks.jar {
-    enabled = true
-}
-
 dependencies {
-    // SpringDoc OpenAPI
-    api("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.13")
+    api(project(":libs:core:exception"))
 
-    // Core exception module for error code references
-    implementation(project(":libs:core:exception"))
-    implementation(project(":libs:api:response"))
-    implementation(project(":libs:contracts:http"))
+    // Logback PatternLayout for PII masking layout
+    compileOnly("ch.qos.logback:logback-classic")
 
     // Lombok
     compileOnly("org.projectlombok:lombok")
@@ -28,4 +17,12 @@ dependencies {
     // Testing
     testImplementation("org.springframework.boot:spring-boot-starter-test")
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+}
+
+tasks.bootJar {
+    enabled = false
+}
+
+tasks.jar {
+    enabled = true
 }
