+++ libs/security/crypto/src/main/java/com/example/security/crypto/encryption/Encryptor.java
@@ -1,0 +1,10 @@
+package com.example.security.crypto.encryption;
+
+public interface Encryptor {
+
+    String encrypt(String plainText);
+
+    String decrypt(String cipherText);
+
+    String getAlgorithm();
+}
