+++ libs/security/crypto/src/main/java/com/example/security/crypto/pii/PiiType.java
@@ -1,0 +1,11 @@
+package com.example.security.crypto.pii;
+
+public enum PiiType {
+    GENERAL,
+    EMAIL,
+    PHONE,
+    NAME,
+    ADDRESS,
+    CARD_NUMBER,
+    SSN
+}
