+++ servers/test/test-server/src/main/java/com/example/testserver/controller/TestController.java
@@ -54,7 +54,8 @@ public class TestController implements
                 Map.of("module", "libs:api:response", "description", "ApiResponse, ErrorResponse, PageResponse"),
                 Map.of("module", "libs:api:exception-handler", "description", "GlobalExceptionHandler"),
                 Map.of("module", "libs:data:entity", "description", "BaseEntity (JPA Auditing + Soft Delete)"),
-                Map.of("module", "libs:security:core", "description", "HMAC-SHA256 서명, 암호화, PII 마스킹"),
+                Map.of("module", "libs:security:context", "description", "AuthContext, HMAC-SHA256 서명 검증, 헤더 파서"),
+                Map.of("module", "libs:security:crypto", "description", "암호화, PII 마스킹"),
                 Map.of("module", "libs:config:kafka", "description", "KafkaTemplate, 멱등성 소비자, DLQ"),
                 Map.of("module", "libs:config:redis", "description", "RedisTemplate, CacheManager"),
                 Map.of("module", "libs:config:resilience", "description", "Circuit Breaker, Retry, Fallback"),
